packages/core/pyproject.toml
[project]
name = "omniscience-core"
version = "0.1.0"
description = "Shared types, config, logging, telemetry for Omniscience"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "prometheus-client>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omniscience_core"]
